fix: address Claude review suggestions — off-by-one, unbounded retry loop, move backoff to utils (#59)
